--- User/includes/landing/footer.php
@@ -1,10 +1,0 @@
-<footer class="container text-center">
-    <!-- <p>&copy; Skincarely.ph &copy; 2020-2021</p> -->
-    <strong>Skincarely.ph</strong> &copy; 2021. All rights reserved.
-</footer>
-<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-<script src="../../dist/js/popper.min.js"></script>
-<script src="../../dist/js/holder.min.js"></script>
-<script src="../../dist/js/bootstrap.min.js"></script>
-</body>
-</html>
