--- User/includes/landing/head.php
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Skincarely | Home</title>
-    <!-- Font awesome icons -->
-    <script src="https://kit.fontawesome.com/d9fb1896a3.js" crossorigin="anonymous"></script>  
-    <!-- Bootstrap -->
-    <link rel="stylesheet" href="dist/css/bootstrap.min.css">
-    <!-- Sripts -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
-    <style>
-        .bg-img {
-            background-image: url('dist/img/landing-cover.jpg');
-            /* background: linear-gradient(0deg, #00000088 30%, #FFFFFF44 100%); */
-
-            height: 100%;
-
-            background-position: center;
-            background-repeat: no-repeat;
-            background-size: cover;
-
-            filter: blur(8px);
-            -webkit-filter: blur(8px);
-            /*
-            Hours wasted trying to make the background appear sa jumbotron: 3
-            If someone knows how to CSS, paggawa ng part na to ty
-            */
-        }
-    </style>
-</head>
-<body style="padding-top: 3.5rem;">
