--- User/includes/login/head.php
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Skincarely | Login</title>
-    <!-- Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <!-- Sripts -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
-    <!-- Google Fonts -->
-    <link href='https://fonts.googleapis.com/css?family=Calistoga' rel='stylesheet'>
-    <link href="https://fonts.googleapis.com/css?family=Poppins" rel="stylesheet" >
-    <!-- Font awesome icons -->
-    <script src="https://kit.fontawesome.com/d9fb1896a3.js" crossorigin="anonymous"></script>
-</head>
-<body>
