+++ User/includes/main/footer.php
@@ -8,7 +8,7 @@
 <!-- jQuery -->
 <script src="plugins/jquery/jquery.min.js"></script>  
 <script src="plugins/bootstrap/js/bootstrap.bundle.min.js"></script>    
-<script src="plugins/bs-custom-file-input/bs-custom-file-input.min.js"></script> 
+<script src="plugins/bs-custom-file-input/bs-custom-file-input.min.js"></script>  
 <!-- Main App -->
 <script src="dist/js/main.js"></script>   
 <script>
