+++ User/includes/main/header.php
@@ -5,10 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>
   Skincarely
-  </title> 
-  <link rel = "icon" href = 
-  "https://i.ibb.co/mJrCGQW/wrench.png" 
-  type = "image/x-icon">    
+  </title>    
   <!-- Bootstrap -->
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
   <link href='https://fonts.googleapis.com/css?family=Calistoga' rel='stylesheet'>
@@ -17,7 +14,7 @@
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css">
   <!-- Ionicons -->
   <link rel="stylesheet" href="https://code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css"> 
-  <!-- Main style -->
+  <!-- Main style --> 
   <link rel="stylesheet" type="text/css"href="dist/css/main.css?version=1"> 
   <!-- overlayScrollbars -->
   <link rel="stylesheet" href="plugins/overlayScrollbars/css/OverlayScrollbars.min.css"> 
@@ -25,4 +22,12 @@
   <script src="https://kit.fontawesome.com/d9fb1896a3.js" crossorigin="anonymous"></script>  
 </head>  
 <body>
-<div class="wrapper">
+<div class="wrapper"> 
+<?php  
+    session_start();
+    if(!isset($_SESSION['loggedin']) || $_SESSION['loggedin'] !== true) {
+        header("location: login_page.php");
+        exit;
+    }
+    include 'includes/main/navbar.php'; 
+?>
