--- User/includes/myaccount/navbar.php
@@ -1,26 +1,0 @@
-<!-- Navbar -->
-<nav class="navbar navbar-expand-md fixed-top navbar-light" style="background-color: #AC7672;">
-    <div class="d-flex align-items-center">
-        <i class="fas fa-crow fa-2x me-3" style="padding: 0 1rem 0;"></i>
-        <a href="#" class="navbar-brand" style="color: #fff;">Skincarely</a>
-    </div>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse" aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon">
-        </span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarCollapse">
-        <ul class="navbar-nav ml-auto">
-            <li class="nav-item">
-                <a href="landing_page.php" class="nav-link" data-widget="control-sidebar" data-controlsidebar-slide="true" role="button">
-                    <i class="fa-solid fa-house" aria-hidden="true"></i>
-                </a>
-            </li>
-            <li class="nav-item">
-                <a href="logout.php" class="nav-link" role="button">
-                    <i class="fa-solid fa-right-from-bracket"></i>
-                </a>
-            </li>
-        </ul>
-    </div>
-</nav>
-<!-- /Navbar -->
