+++ User/logout.php
@@ -1,10 +1,6 @@
 <?php
-
-session_start();
-
-$_SESSION = array();
-
-session_destroy();
-header("location: login_page.php");
-
+    session_start();
+    $_SESSION = array();
+    session_destroy();
+    header("location: login_page.php");
 ?>
